requirements.txt
# Database
psycopg2-binary==2.9.9

# Data handling
pandas==2.1.4
python-dotenv==1.0.0

# Date/time
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0